requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap4==1.1.1
django-extensions==2.2.9
django-js-asset==1.2.2
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-popup-view-field==0.6.1
django-redis-sessions==0.6.1
graphviz==0.14
gunicorn==20.0.4
Pillow==7.1.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
pydot==1.4.1
pyparsing==2.4.7
python-slugify==4.0.0
pytz==2019.3
redis==3.5.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.0.1
